fix menu error

+++ Assets/Scripts/HealthBar.cs
@@ -12,7 +12,7 @@ public class HealthBar : MonoBehaviour
 
 	public void SetMaxHealth(int health)
 	{
-		slider.maxValue = health;
+		slider.maxValue = (float)health;
 		slider.value = health;
 
 		fill.color = gradient.Evaluate(1f);
